requirements.txt
# Core Data Processing
pyspark==3.3.2
pandas==1.5.3
numpy==1.24.3

# Google Cloud Platform
google-cloud-storage==2.10.0
google-cloud-bigquery==3.11.4
google-cloud-dataproc==5.4.3

# Hadoop/Hive Integration
pyhive==0.7.0
thrift==0.16.0
sasl==0.3.1

# Data Analysis
scipy==1.10.1
scikit-learn==1.2.2
statsmodels==0.14.0

# Visualization
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.14.1

# Workflow Orchestration
apache-airflow==2.6.3

# Database
sqlalchemy==2.0.15

# Utilities
python-dotenv==1.0.0
pyyaml==6.0
click==8.1.3

# Testing
pytest==7.3.1
pytest-cov==4.1.0

# Jupyter
jupyter==1.0.0
jupyterlab==4.0.2
ipykernel==6.23.1

# Code Quality
black==23.3.0
flake8==6.0.0
pylint==2.17.4
